added sfx + vfx + achievements system + settings

+++ Eggs Factory Idle/Assets/Game/Scripts/Interfaces/ICollectable.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 public interface ICollectable
 {
     void Collect();
-    int GetCoinReward();
+    float GetCoinReward();
 }
